remove ClassHasProperty
