--- src/Framework/Constraint/Object/ClassHasProperty.php
@@ -1,85 +1,0 @@
-<?php declare(strict_types=1);
-/*
- * This file is part of PHPUnit.
- *
- * (c) Sebastian Bergmann <sebastian@phpunit.de>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-namespace PHPUnit\Framework\Constraint;
-
-use function get_class;
-use function is_object;
-use function sprintf;
-use PHPUnit\Framework\Exception;
-use ReflectionClass;
-use ReflectionException;
-
-/**
- * @no-named-arguments Parameter names are not covered by the backward compatibility promise for PHPUnit
- */
-class ClassHasProperty extends Constraint
-{
-    private readonly string $propertyName;
-
-    public function __construct(string $propertyName)
-    {
-        $this->propertyName = $propertyName;
-    }
-
-    /**
-     * Returns a string representation of the constraint.
-     */
-    public function toString(): string
-    {
-        return sprintf(
-            'has property "%s"',
-            $this->propertyName
-        );
-    }
-
-    /**
-     * Evaluates the constraint for parameter $other. Returns true if the
-     * constraint is met, false otherwise.
-     *
-     * @param mixed $other value or object to evaluate
-     */
-    protected function matches(mixed $other): bool
-    {
-        try {
-            return (new ReflectionClass($other))->hasProperty($this->propertyName);
-            // @codeCoverageIgnoreStart
-        } catch (ReflectionException $e) {
-            throw new Exception(
-                $e->getMessage(),
-                $e->getCode(),
-                $e
-            );
-        }
-        // @codeCoverageIgnoreEnd
-    }
-
-    /**
-     * Returns the description of the failure.
-     *
-     * The beginning of failure messages is "Failed asserting that" in most
-     * cases. This method should return the second part of that sentence.
-     *
-     * @param mixed $other evaluated value or object
-     */
-    protected function failureDescription(mixed $other): string
-    {
-        return sprintf(
-            '%sclass "%s" %s',
-            is_object($other) ? 'object of ' : '',
-            is_object($other) ? get_class($other) : $other,
-            $this->toString()
-        );
-    }
-
-    protected function propertyName(): string
-    {
-        return $this->propertyName;
-    }
-}
