+++ src/Framework/Constraint/Object/ObjectHasProperty.php
@@ -14,16 +14,52 @@ use ReflectionObject;
 /**
  * @no-named-arguments Parameter names are not covered by the backward compatibility promise for PHPUnit
  */
-final class ObjectHasProperty extends ClassHasProperty
+final class ObjectHasProperty extends Constraint
 {
+    private readonly string $propertyName;
+
+    public function __construct(string $propertyName)
+    {
+        $this->propertyName = $propertyName;
+    }
+
+    /**
+     * Returns a string representation of the constraint.
+     */
+    public function toString(): string
+    {
+        return sprintf(
+            'has property "%s"',
+            $this->propertyName
+        );
+    }
+
     /**
      * Evaluates the constraint for parameter $other. Returns true if the
      * constraint is met, false otherwise.
      *
      * @param mixed $other value or object to evaluate
      */
-    protected function matches($other): bool
+    protected function matches(mixed $other): bool
     {
-        return (new ReflectionObject($other))->hasProperty($this->propertyName());
+        return (new ReflectionObject($other))->hasProperty($this->propertyName);
+    }
+
+    /**
+     * Returns the description of the failure.
+     *
+     * The beginning of failure messages is "Failed asserting that" in most
+     * cases. This method should return the second part of that sentence.
+     *
+     * @param mixed $other evaluated value or object
+     */
+    protected function failureDescription(mixed $other): string
+    {
+        return sprintf(
+            '%sclass "%s" %s',
+            is_object($other) ? 'object of ' : '',
+            is_object($other) ? get_class($other) : $other,
+            $this->toString()
+        );
     }
 }
